+++ examples/data-sources/catalystcenter_role/data-source.tf
@@ -1,0 +1,3 @@
+data "catalystcenter_role" "example" {
+  id = "Role1"
+}