+++ examples/resources/catalystcenter_role/import.sh
@@ -1,0 +1,1 @@
+terraform import catalystcenter_role.example "Role1"