+++ examples/resources/catalystcenter_role/resource.tf
@@ -1,0 +1,10 @@
+resource "catalystcenter_role" "example" {
+  name        = "Role1"
+  description = "Role1 description"
+  resource_types = [
+    {
+      type       = "Platform"
+      operations = ["gRead"]
+    }
+  ]
+}