+++ internal/provider/model_catalystcenter_credentials_https_read.go
@@ -51,8 +51,8 @@ func (data CredentialsHTTPSRead) getPath() string {
 //template:begin toBody
 func (data CredentialsHTTPSRead) toBody(ctx context.Context, state CredentialsHTTPSRead) string {
 	body := ""
-	if data.Id.ValueString() != "" {
-		body, _ = sjson.Set(body, "httpsRead.0.id", data.Id.ValueString())
+	if state.Id.ValueString() != "" {
+		body, _ = sjson.Set(body, "httpsRead.0.id", state.Id.ValueString())
 	}
 	if !data.Description.IsNull() {
 		body, _ = sjson.Set(body, "httpsRead.0.description", data.Description.ValueString())