+++ internal/provider/model_catalystcenter_project.go
@@ -49,8 +49,8 @@ func (data Project) getPath() string {
 //template:begin toBody
 func (data Project) toBody(ctx context.Context, state Project) string {
 	body := ""
-	if data.Id.ValueString() != "" {
-		body, _ = sjson.Set(body, "id", data.Id.ValueString())
+	if state.Id.ValueString() != "" {
+		body, _ = sjson.Set(body, "id", state.Id.ValueString())
 	}
 	if !data.Name.IsNull() {
 		body, _ = sjson.Set(body, "name", data.Name.ValueString())