+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - BREAKING CHANGE: Rename `catalystcenter_device` resource and data source to `catalystcenter_pnp_device`
 - BREAKING CHANGE: Rename `catalystcenter_device_claim_site` resource and data source to `catalystcenter_pnp_device_claim_site`
 - Add `catalystcenter_device` resource
+- Add `catalystcenter_role` resource and data source
 
 ## 0.1.2